requirements.txt
PyYAML==6.0
pandas==2.0.1
os-sys==2.1.4
mysql-connector-python==8.0.33
MySQL-python==1.2.5
mysql-connector==2.2.9
Flask==2.3.2
